requirements.txt
simplejson>=3.3.0
lxml==3.0.2
zope.interface>=4.0.5
twisted==12.3.0